requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
googletrans==4.0.0-rc1
python-docx
openpyxl
xlrd
PyPDF2
